requirements.txt
altgraph==0.17.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
fonttools==4.53.0
glfw==2.7.0
idna==3.7
ifaddr==0.2.0
importlib_metadata==7.1.0
ipydatawidgets==4.3.5
ipython==8.25.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
open3d==0.18.0
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.0
parso==0.8.4
pefile==2024.8.26
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
plyfile==1.0.3
prompt_toolkit==3.0.45
pure-eval==0.2.2
pyassimp==5.2.5
pybullet==3.2.6
pyfbx==0.0.8
pygame==2.6.0
pyglet==2.0.15
Pygments==2.18.0
pyinstaller-hooks-contrib==2024.8
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pythreejs==2.4.2
PyWavefront==1.3.3
pywin32==306
pywin32-ctypes==0.2.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.18.1
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
tenacity==8.3.0
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
zeroconf==0.132.2
zipp==3.19.1
